// src/types/ContactType.ts
export type ContactDatasType= {
    title: string;
    introText: string;
    labelName: string;
    labelSurname: string;
    labelSociety: string;
    labelEmail: string;
    labelMessage: string;
    placeHolder: string;
    labelSubmit: string;
    formMissedField: string;
    formWrongEmail: string;
    formSendEmail: string;
    formErrorEmail: string;
};
export type FormContactType={
    name: string;
    surname: string;
    email: string;
    society: string;
    message: string;
}